cms_project/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
coverage==7.4.0
pytest==7.4.4
pytest-django==4.7.0
drf-yasg==1.21.7
django-filter==23.2
